Update README.md
An airline reservation system in C++ is a program that allows users to reserve seats on flights. The system is designed to be user-friendly, with a simple and intuitive interface. The program consists of several modules, including a  manual reservation module, packages module and a payment module. The manual reservation module allows users to search for available flights, select seats, and make reservations